requirements.txt
streamlit>=1.27.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
plotly>=5.15.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
requests>=2.31.0
pillow>=10.0.0
